+++ settings.py
@@ -2,7 +2,7 @@
 
 import os
 
-BOT_VERSION = "6.0.0"
+BOT_VERSION = "6.1.0"
 BACKUP_DIR = 'db_backups'
 
 def get_base_dir():