+++ utils.py
@@ -1,5 +1,7 @@
 # -*- coding: utf-8 -*-
 
+# utils.py
+
 import discord
 import logging